requirements.txt
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
python-telegram-bot>=20.0
telethon>=1.34.0
